actualizado los datos y ejercicios de la practica de funciones en javascript
