+++ javaScript/scripts-classPractice/document-element.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">   
     <title>Document</title>
+    
      <style>
         #parrafo{
             border: 1px solid red;
@@ -16,7 +17,7 @@
 
     <button onclick="alert('Hiciste click en el botón')">Click</button>
 
-    <!-- <img src="hola.png" id="miImagen"> -->
+    <!-- <img src="hola.png" id="miImagen" alt="deberia haber una imagen"> -->
 
     <div id="cualquier_cosa" class="div_transparente texto">
         Este es un Div
@@ -42,12 +43,7 @@
 
     <input type="number"  id="number1" placeholder="Ingrese primer número">
     <input type="number" id="number2" placeholder="Ingrese segundo número">
-    <button type="button" id="boton3">Suma</button>
-
-
-
-
-
+    <button type="button" id="boton3">Suma</button>   
     <script type="text/javascript" src="./document-element.js"></script>
